==> 02-predicting-molecular-properties/notebooks/01-predicting-molecular-properties.ipynb <==
* This notebook aims to predict the `scalar_coupling_constants` between atom pairs in molecules,given the two atom types (e.g., C and H), the coupling type (e.g., 2JHC).

# importing the libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os
os.chdir(r"D:\Random_Datasets\Kaggle\champs-scalar-coupling")
import xgboost as xgb
import pickle
from sklearn.model_selection import train_test_split

# 1. Importing the Datasets

# importing the dataset
train_df = pd.read_csv("train.csv")
test_df = pd.read_csv("test.csv")
dipole_df = pd.read_csv("dipole_moments.csv")
magnetic_shielding_tensors_df = pd.read_csv("magnetic_shielding_tensors.csv")
potential_energy_df = pd.read_csv("potential_energy.csv")
structures_df = pd.read_csv("structures.csv")

# printing the size of datasets
print("Length of train_df = {}".format(len(train_df)))
print("Length of test_df = {}".format(len(test_df)))
print("Length of dipole_df = {}".format(len(dipole_df)))
print("Length of magnetic_shielding_tensors_df = {}".format(len(magnetic_shielding_tensors_df)))
print("Length of potential_energy_df = {}".format(len(potential_energy_df)))
print("Length of structures_df = {}".format(len(structures_df)))

### 1.1 train_df
- *molecule_name* = name of the molecule where the coupling constant originates. The corresponding values of XYZ are there in **structures.df**.
- *atom_index_0*, *atom_index_1* = atom pair creating coupling constant.
- *type* = coupling type
- *scaler_coupling_constant* = **predictor**

train_df.head() 

test_df.head()

def info_df(data):
    """
    returns the dataset of unique and missing values
    """
    return pd.DataFrame({
        "unique_values": data.nunique(),
        "null_values": data.isnull().sum()
    })

train_info = info_df(train_df)
train_info

test_info = info_df(test_df)
test_info

### dipole_df
Contains the molecular electric dipole moments. These are three dimensional vectors that indicate the charge distribution in the molecule.
X, Y, Z are three components of dipole moment.

dipole_df.head()

dipole_info = info_df(dipole_df)
dipole_info

The `dipole_df` contains 85003 values, unique for each molecule. This can merged with the main dataframe to increase the predictors.

## magnetic_shielding_tensors_df
contains the magnetic shielding tensors for all atoms in the molecules. The first column `(molecule_name)` contains the molecule name, the second column `(atom_index)` contains the index of the atom in the molecule, the third to eleventh columns contain the `XX`, `YX`, `ZX`, `XY`, `YY`, `ZY`, `XZ`, `YZ` and `ZZ` elements of the tensor/matrix respectively.

magnetic_shielding_tensors_df.head()

mst_info_df = info_df(magnetic_shielding_tensors_df)
mst_info_df

## Note
How to use atom index in molecule name as only one index is given?

## potential_energy_df
Contains the potential energy of the molecules. The first column `(molecule_name)` contains the name of the molecule, the second column `(potential_energy)` contains the potential energy of the molecule.

potential_energy_df.head()

pe_info_df = info_df(potential_energy_df)
pe_info_df

## sturctures_df
X, Y and Z  = cartesian coordinates

structures_df.sample(5)

structures_info_df = info_df(structures_df)
structures_info_df

---
# 2. Feature Engineering

## 2.1 Merging the datasets

def map_atom_info(df, atom_idx):
    """
    works in three steps :-
    1. merges train_df/test_df with structures_df
    2. drops the atom_index on which it is merged
    3. renames the columns
    """
    df = pd.merge(df, structures_df, how = 'left',
                  left_on  = ['molecule_name', f'atom_index_{atom_idx}'],
                  right_on = ['molecule_name',  'atom_index'])
    
    df = df.drop('atom_index', axis=1)
    df = df.rename(columns={'atom': f'atom_{atom_idx}',
                            'x': f'x_{atom_idx}',
                            'y': f'y_{atom_idx}',
                            'z': f'z_{atom_idx}'})
    return df


# implementing above function
train_df = map_atom_info(train_df, 0)
train_df = map_atom_info(train_df, 1)

test_df = map_atom_info(test_df, 0)
test_df = map_atom_info(test_df, 1)

train_df.sample(4)

test_df.sample(4)

### Columns after merging the datasets :-
- molecule name
- atom_index_0
- atom_index_1
- type = presumably scaling coupling types. `T` in evaluation formula
- scalar_coupling_constant = **Predictor**
- atom_0
- x_0, y_0, z_0 = cartesian places of atom_0
- atom_1
- x_1, y_1, z_1 = cartesian places of atom_1

## 2.2 Distances

####################################################
#### Magnitude of distance b/w two atoms ###########
####################################################

def dist_magnitude(df):
    x = df["x_0"] - df["x_1"]
    y = df["y_0"] - df["y_1"]
    z = df["z_0"] - df["z_1"]
    
    dist = (x**2 + y**2 + z**2)**0.5
    
    return dist

train_df["abs_distance"] = dist_magnitude(train_df)
test_df["abs_distance"] = dist_magnitude(test_df)

###################################################
### individual cartesian distance #################
###################################################

def dist_individual(df):
    df["x_dist"] = df["x_0"] - df["x_1"]
    df["y_dist"] = df["y_0"] - df["y_1"]
    df["z_dist"] = df["z_0"] - df["z_1"]
    
    return df

train_df = dist_individual(train_df)
test_df = dist_individual(test_df)

train_df.head()

test_df.head()

## 2.3 Descriptive Statistics features
These features will be decided on the `type` column. The features values created in train_df will used in the testset to make the model more robust. The following features will be created :-
- mean_value
- difference from the mean
- std_dev
- z_score

train_info = info_df(train_df)
train_info

features_df = train_df.groupby("type").agg({
    "x_0": ["mean", "median", "max", "min", "std"],
    "y_0": ["mean", "median", "max", "min", "std"],
    "z_0": ["mean", "median", "max", "min", "std"],
    "x_1": ["mean", "median", "max", "min", "std"],
    "y_1": ["mean", "median", "max", "min", "std"],
    "z_1": ["mean", "median", "max", "min", "std"],
    "x_dist": ["mean", "median", "max", "min", "std"],
    "y_dist": ["mean", "median", "max", "min", "std"],
    "z_dist": ["mean", "median", "max", "min", "std"],
    "abs_distance": ["mean", "median", "max", "min", "std"]
})

features_df.columns = ["x_0_mean", "x_0_median", "x_0_max", "x_0_min", "x_0_std", 
                       "y_0_mean", "y_0_median", "y_0_max", "y_0_min", "y_0_std",
                       "z_0_mean", "z_0_median", "z_0_max", "z_0_min", "z_0_std",
                       "abs_distance_mean", "abs_distance_median", "abs_distance_max", "abs_distance_min", "abs_distance_std",
                       "x_dist_mean", "x_dist_median", "x_dist_max", "x_dist_min", "x_dist_std",
                       "y_dist_mean", "y_dist_median", "y_dist_max", "y_dist_min", "y_dist_std",
                       "z_dist_mean", "z_dist_median", "z_dist_max", "z_dist_min", "z_dist_std",
                       "x_1_mean", "x_1_median", "x_1_max", "x_1_min", "x_1_std", 
                       "y_1_mean", "y_1_median", "y_1_max", "y_1_min", "y_1_std",
                       "z_1_mean", "z_1_median", "z_1_max", "z_1_min", "z_1_std"
                      ]

features_df = features_df.reset_index()

features_df

train_df = pd.merge(train_df, features_df, on = "type", how = "inner")
test_df = pd.merge(test_df, features_df, on = "type", how = "inner")

train_df.sample(4)

---
## 3. Model 01 - XGBoost

train_df["type"] = train_df["type"].map({
    '1JHC': 0, '2JHH': 1, '1JHN': 2, '2JHN': 3, '2JHC': 4, '3JHH': 5, '3JHC': 6, '3JHN': 7
})

train_df["atom_0"] = train_df["atom_0"].map({
    "H": 0
})

train_df["atom_1"] = train_df["atom_1"].map({
    'C': 0, 'H': 1, 'N': 0
})

X = train_df.drop(columns = ["id", "molecule_name", "scalar_coupling_constant"])
y = train_df["scalar_coupling_constant"]

x_train, x_valid, y_train, y_valid = train_test_split(X, y, test_size = 0.2, random_state = 42)

# preprocessing test_df
id_ = test_df["id"]
test_df = test_df.drop(columns = ["id", "molecule_name"])

test_df["type"] = test_df["type"].map({
    '1JHC': 0, '2JHH': 1, '1JHN': 2, '2JHN': 3, '2JHC': 4, '3JHH': 5, '3JHC': 6, '3JHN': 7
})

test_df["atom_0"] = test_df["atom_0"].map({
    "H": 0
})

test_df["atom_1"] = test_df["atom_1"].map({
    'C': 0, 'H': 1, 'N': 0
})

train_df.to_csv("train_2.csv")
test_df.to_csv("test_2.csv")

print("x_train shape = {}".format(x_train.shape))
print("x_valid shape = {}".format(x_valid.shape))
print("y_train shape = {}".format(y_train.shape))
print("y_valid shape = {}".format(y_valid.shape))

# converting the dataset into DMatrices
dtrain = xgb.DMatrix(x_train, label = y_train)
dvalid = xgb.DMatrix(x_valid, label = y_valid)

# training the Model
params = {
    "eta":0.1,
    "alpha": 0.1,
    "max_leaves": 128,
    "max_depth": 9,
    "n_estimators": 100,
    "max_depth": 6,
    "subsample": 0.8,
    "colsample_bytree": 1,
    "gamma": 0,
    "eval_metric": "mae",
    "nthreads": 4,
    "objective": "reg:linear",
    "silent": 1
}

model_train = xgb.train(params, dtrain, 5000, evals = [(dvalid, "valid_set")], verbose_eval=10, early_stopping_rounds = 40)

pickle.dump(model_train, open('pmp_xgb.pickle.dat','wb'))

dtest = xgb.DMatrix(test_df)

y_pred = model_train.predict(dtest)

result_df = pd.DataFrame({
    "id": id_,
    "scalar_coupling_constant": y_pred
})

result_df.sample(10)

result_df.to_csv("submission.csv", index = False)
