==> 02-predicting-molecular-properties/notebooks/02-predicting-molecular-properties.ipynb <==
# importing the libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import os
os.chdir(r"D:\Random_Datasets\Kaggle\champs-scalar-coupling")
import xgboost as xgb
import pickle
from sklearn.model_selection import train_test_split

pd.set_option("display.max_columns", 100)
pd.set_option("display.max_rows", 50)

%%time
train_df = pd.read_csv("train_2.csv")
test_df = pd.read_csv("test_2.csv")

test_df.head()

%%time
train_df = train_df.drop(columns = ["Unnamed: 0", "id", "molecule_name"])

train_df.head()

test_df = test_df.drop(columns = ["Unnamed: 0"])
test_df.head()

print("train_df = {}".format(train_df.shape))
print("test_df = {}".format(test_df.shape))

## Feature Creation

We will expand our feature set by creating more features based on distances. Following are the features, we will create for next iteration of the model training:
1. *Statistical summary differences* - Subtract x_0 from the derived features. The new feature created will be x_0_mean_diff, x_0_kurtosis_diff, x_0_max_diff, x_0_min_diff and x_0_std_diff. Delete the parent columns afterwards.
2. Repeat this exercise for
    - for other atomic features like y_0, z_0 etc.
    - for abs_distance
    - for x_dist, y_dist, z_dist
3. Angle Based features

def diff_columns(dataframe, col_family):
    """
    this functions takes the dataframe and col_family as input and return the dataframe
    where differenced features are returned.
    """
    df = dataframe
    col_list = [col for col in df.columns if col.startswith(col_family)]
    
    for i in col_list[1:]:
        df[i + "_diff"] = df[i] - df[col_list[0]]
        
    return df

%%time
col_family = ["x_0", "y_0", "z_0", "x_1", "y_1", "z_1", "abs_distance", "x_dist", "y_dist", "z_dist"]

for i in col_family:
    train_df = diff_columns(train_df, i)

train_df.head()

train_df.shape

# similarly implementing for test_df
for i in col_family:
    test_df = diff_columns(test_df, i)
    
test_df.shape

%%time
drop_cols = [col for col in train_df.columns if col.endswith("mean") or col.endswith("median") or col.endswith("max") or 
             col.endswith("min") or col.endswith("std")]

cols = [col for col in train_df.columns if col not in drop_cols]
train_df = train_df[cols]

# removing target columns
cols.remove("scalar_coupling_constant")
test_df = test_df[cols]

print("train_df = {}".format(train_df.shape))
print("test_df = {}".format(test_df.shape))

test = pd.read_csv("test.csv")
id_ = list(test["id"])
test_df.insert(0, "id", id_)

%%time
train_df.to_csv("train_3.csv", index = False)
test_df.to_csv("test_3.csv", index = False)

## Train the Model

%%time
x_train, x_valid, y_train, y_valid = train_test_split(train_df.drop(columns = ["scalar_coupling_constant"]), 
                                                      train_df["scalar_coupling_constant"], 
                                                      test_size = 0.2, 
                                                      random_state = 42)

%%time
# converting the dataset into DMatrices
dtrain = xgb.DMatrix(x_train, label = y_train)
dvalid = xgb.DMatrix(x_valid, label = y_valid)

%%time
# training the Model
params = {
    "eta":0.1,
    "alpha": 0.1,
    "max_leaves": 128,
    "max_depth": 9,
    "n_estimators": 100,
    "max_depth": 6,
    "subsample": 0.8,
    "colsample_bytree": 1,
    "gamma": 0,
    "eval_metric": "mae",
    "nthreads": 4,
    "objective": "reg:linear",
    "silent": 1
}

model_train = xgb.train(params, dtrain, 2500, evals = [(dvalid, "valid_set")], verbose_eval = 100, early_stopping_rounds = 40)

pickle.dump(model_train, open('pmp_xgb_2.pickle.dat','wb'))

dtest = xgb.DMatrix(test_df)

%%time
y_pred = model_train.predict(dtest)

result_df = pd.DataFrame({
    "id": id_,
    "scalar_coupling_constant": y_pred
})

result_df.sample(10)

result_df.to_csv("submission_02.csv", index = False)
