==> 02-predicting-molecular-properties/notebooks/03-feature-engineering-with-dask.ipynb <==
# Setting up the environment

# importing the libraries

import os
import numpy as np
import pandas as pd
import dask.dataframe as dd
import matplotlib.pyplot as plt
import seaborn as sns

# setting up the parameters

root_dir = os.path.dirname(os.path.abspath(os.getcwd()))
pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", 10)
plt.rcParams["figure.figsize"] = (16, 8)
sns.set_style("darkgrid")

---
# 1. Importing the Data

%%time
# importing the dataset

train_df = pd.read_csv(os.path.join(root_dir, "data", "train_3.csv"))
test_df = pd.read_csv(os.path.join(root_dir, "data", "test_3.csv"))

# 2. Machine Learning
## 2.1. DNN Regressor
- Seperate the label and variables
- MinMaxScale the numerical columns
- Convert the raw dataset into Dense tensors
- 

%%time
train_df.head()

x_train = train_df.drop(columns = ["scalar_coupling_constant"])     # predictors
y_train = train_df["scalar_coupling_constant"]                     # labels

del train_df

x_train.info()

# all the values defined above as int are categorical columns.
boo = type(np.array(x_train["atom_index_0"])[1])

categorical_columns = [i for i in x_train.columns if type(np.array(x_train[i])[1]) == boo]
continuous_columns = [i for i in x_train.columns if i not in categorical_columns]

from sklearn.preprocessing import MinMaxScalar

# scaling the continuous columns
mms =
